fix: update 'a' behavior in select mode
() represents position of block cursor in normal/select mode, | is position of line cursor in insert mode

with the helix preset starting in normal mode, given: `(x)y` I expect `v → a` to result in `xy|` but it instead results in `x|y`. 

This appears to be because `cursorRight` while there's a selection cancels the selection without actually updating the position of the cursor.